+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changes since the last release
 
 * Added "Set as Target" in Monitor (HaullyGames)
+* Connection: connection rate is minimum rate in ControlPath. (HaullyGames)
 ------------------------------------------------------------------------------------------------------
 
 ## v2.0.0 for KSP 1.5.x , 1.4.x and 1.3.1